+++ src/parser/tlf.rs
@@ -4,11 +4,15 @@ use core::fmt;
 
 use crate::parser::ParseError;
 
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
 use super::{take_byte, SmlParse};
 
 use super::ResTy;
 
 /// Error type used when parsing a `TypeLengthField`
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum TlfParseError {
     /// The length field of a TLF overflowed